# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "La bonne affaire",
        "description": "",
        "contact": {
            "email": "nathalietaffot@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User",
                "description": "Get User",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "affichage d'un utilisateur."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "description": "Update User",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "reponse de la modification"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/send-email": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "send email for verify",
                "description": "verify email",
                "operationId": "sendEmail",
                "requestBody": {
                    "description": "Verification de l'email d'un utilisateur",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "votre email",
                                        "type": "string",
                                        "format": "string",
                                        "example": "examples@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Verification de l'email."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users/avatar": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "upload avatar file",
                "description": "upload avatar file",
                "operationId": "uploadAvatar",
                "requestBody": {
                    "description": "Telechargement du profil utilisateur",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "description": "votre photos de profil",
                                        "type": "file",
                                        "format": "image",
                                        "example": "https://image.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "avatar_path": {
                                            "type": "string",
                                            "example": "users/avatar/ghRfjbiJHOvnMBaeerGTwCbYxV0WEnRuRPFod9N3.jpg"
                                        },
                                        "avatar_url": {
                                            "type": "string",
                                            "example": "http://labonneaffaire.test/users/avatar/ghRfjbiJHOvnMBaeerGTwCbYxV0WEnRuRPFod9N3.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}